ChromecastSessionManager: Set source startTime when changing techs
When we are switching to ChromeCast, we want to have it start playback
at currentTime.

Otherwise, it would need to buffer the media twice, once at 00:00 and
once again for currentTime.

Also, the current implementation to call player.currentTime on
player.ready doesn't seem to work well with the ChromeCast tech as
the seek instruction is sent too early and is ignored by the ChromeCast.